Overhauled the way Datasources are managed and created.
Datasources are now created from items that can be called. They are lazy setup the first time they are called. The means function, anonymous functions, or when classes are called as a function (__invoke). It should return an instanciated datasource that is then cached by the manager.

The arguments passed to the factory that setups up the datasource are $params (a key/value array of params), $name, and $manager (the datasource manager).

For examples of how to create a datasource see the mongo and PDO examples. The mongo one shows an example of how the datasource can interact with a manager to add an additional datasource.

To the manager and context new functions for adding and removing a datasource were added. Default datasources were removed and instead should be replaced by a datasource that manages any master/slaves and manage it's own defaults.